Updated postcard with changes from upvotes
